Working on AbstractManager and GLFWManager
Committing my progress so far.
There is a distinct possibility that I might have to undo all work done for the GLFWManager and AbstractManager so far, but I am hoping this is not the case.

As to what I have done, I am trying to maintain as much of the code that is already in place, just moving it around and tweaking tiny things.